Add generateLunaPayload function to send luna message to DB8
:Release Notes:
Add generateLunaPayload function to send luna message to DB8

:Detailed Notes:
Add generateLunaPayload function to send luna message to DB8.
Add getImageList and getImageMetadata response.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-114194] Implement media indexer client library
[PLAT-115211] Implement image list and metadata response in media
              indexer client

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediaindexer/+/280350
Tested-by: Jimmy Ohn <yongjin.ohn@lge.com>

+++ test/mediaindexerclient/MediaIndexerClientTest.cpp
@@ -110,7 +110,7 @@ static bool processCommand(const std::string& cmd, MediaIndexerClient &client, G
 static void my_indexer_callback(MediaIndexerClientEvent event, void* clientData, void* userData)
 {
     std::cout << "my_indexer_callback" << std::endl;
-    std::cout << "event : " << event << std::endl;
+    std::cout << "event : " << static_cast<int>(event) << std::endl;
     std::cout << "thread[" << std::this_thread::get_id() << "]" << std::endl;
 }
